+++ Server/DataAndAssistance/RelativeDatabase.h
@@ -21,11 +21,11 @@ public:
 
     void setClassifiedRelatives(vector<ClassifiedRelativeVector *> classifiedRelativesVec);
 
-    void setKElement(int k);
+    void setKElement(unsigned long k);
 
     void setMetric(string metric);
 
-    int getKElement();
+    unsigned long getKElement();
 
     string getMetric();
 
@@ -33,7 +33,7 @@ private:
 
     static RelativeDatabase *relativeDatabasePtr;
 
-    int kElement = 5;
+    unsigned long kElement = 5;
 
     string distanceMetric = "AUC";
